backend/requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
coverage==7.8.0
cryptography==41.0.0
defusedxml==0.7.1
Django==5.1.7
django-cors-headers==4.7.0
django-crispy-forms==1.14.0
django-filter==25.1
django-multiselectfield==0.1.13
django-mysql==4.16.0
django-rest-knox==5.0.2
djangorestframework==3.15.2
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
httplib2==0.22.0
idna==3.10
oauthlib==3.2.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
typing_extensions==4.13.0
urllib3==2.3.0
